+++ CotEditor/Sources/CEConsolePanelController.h
@@ -9,7 +9,7 @@
  encoding="UTF-8"
  ------------------------------------------------------------------------------
  
- © 2014 CotEditor Project
+ © 2014-2015 1024jp
  
  This program is free software; you can redistribute it and/or modify it under
  the terms of the GNU General Public License as published by the Free Software
@@ -32,13 +32,13 @@
 
 @interface CEConsolePanelController : CEPanelController
 
-- (void)addErrorString:(NSString *)string;
+- (void)appendMessage:(NSString *)message title:(NSString *)title;
 
 @end
 
 
 
 
-@interface CEScriptErrorView : NSTextView
+@interface CEConsoleTextView : NSTextView
 
 @end